v0.4.2
Removed the date-fns dependency. The class now uses string parsing to interpret dates and provides a customization option through 'parseDate', allowing users to adjust the behavior if necessary. This change reduces the library's size and removes the reliance on date-fns, making it compatible with Node.js versions that do not yet support Intl.